+++ docs/user_guide/changelog.rst
@@ -14,6 +14,7 @@ New features
 - Add support for PopART name scopes via ``poptorch.NameScope``
 - Add mixed precision automatic casting
 - Add support for ``torch.cross``
+- Add support for ``torch.median``
 
 v2.0 (Poplar SDK 2.0)
 =====================